+++ src/engine/sparrowsprite.hpp
@@ -19,7 +19,9 @@ class SparrowSprite : public AnimatedSprite {
     SparrowSprite(float x, float y);
     ~SparrowSprite() override;
     void loadGraphic(const std::string &imagePath, const std::string &xmlPath);
-    void addAnimation(const std::string &name, const std::string &prefix, int framerate, std::vector<uint8_t> indices = {}, bool looped = false);
+    void addAnimation(const std::string &name, const std::string &prefix,
+                      uint8_t framerate, std::vector<uint8_t> indices = {},
+                      bool looped = false);
     void update(float delta) override;
 
     void draw(float x, float y) override;
